requirements.txt
torch>=1.10.0
transformers>=4.15.0
pandas>=1.3.0
numpy>=1.20.0
datasets>=1.18.0
gradio>=3.0.0
evaluate
rouge_score>=0.1.0
nltk>=3.6.0
google-generativeai>=0.1.0
python-dotenv>=0.19.0
accelerate>=0.26.0